# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de Agenda",
        "description": "L5 Swagger",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/eventos": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar todos los eventos del usuario autenticado",
                "operationId": "8df3c75a189cf773a3eb105ea46e145a",
                "responses": {
                    "200": {
                        "description": "Lista de eventos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Evento"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Crear un nuevo evento",
                "operationId": "4594ce682158c01a68019d32f26180e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "fecha_inicio"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Cumpleaños"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Fiesta sorpresa"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-30"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-30"
                                    },
                                    "etiqueta": {
                                        "type": "string",
                                        "example": "Personal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Evento"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/eventos/{id}": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Ver detalle de un evento",
                "operationId": "09ec864fa45a38fe55c3078da04e8a13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del evento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Evento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Actualizar un evento",
                "operationId": "032b9536863e8e3b817a73e44d058c1d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del evento a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Cumpleaños actualizado"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Fiesta sorpresa con amigos"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-30"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-30"
                                    },
                                    "etiqueta": {
                                        "type": "string",
                                        "example": "Trabajo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Evento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Eliminar un evento",
                "operationId": "a904dc7b49560c21fe21f57e1df8813a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del evento a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Evento eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Evento": {
                "title": "Evento",
                "properties": {
                    "id": {
                        "description": "* @OA\\Schema(\n*     schema=\"Evento\",\n*     type=\"object\",\n*     title=\"Evento\",\n*     @OA\\Property(property=\"id\", type=\"integer\", example=1),\n*     @OA\\Property(property=\"user_id\", type=\"integer\", example=1),\n*     @OA\\Property(property=\"titulo\", type=\"string\", example=\"Cumpleaños\"),\n*     @OA\\Property(property=\"descripcion\", type=\"string\", example=\"Fiesta sorpresa\"),\n*     @OA\\Property(property=\"fecha_inicio\", type=\"string\", format=\"date\", example=\"2025-11-30\"),\n*     @OA\\Property(property=\"fecha_fin\", type=\"string\", format=\"date\", example=\"2025-11-30\"),\n*     @OA\\Property(property=\"etiqueta\", type=\"string\", example=\"Personal\"),\n*     @OA\\Property(property=\"created_at\", type=\"string\", format=\"date-time\"),\n*     @OA\\Property(property=\"updated_at\", type=\"string\", format=\"date-time\")\n* )",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Cumpleaños"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Fiesta sorpresa"
                    },
                    "fecha_inicio": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-30"
                    },
                    "fecha_fin": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-30"
                    },
                    "etiqueta": {
                        "type": "string",
                        "example": "Personal"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}